# CUDA full GPU acceleration, KV cache in VRAM