This PR adds GPU acceleration for all remaining ggml tensors that didn't yet have it. Especially for long generations this makes a large difference because the KV cache is still CPU only on master and gets larger as the context fills up. Prompt processing is also significantly faster because the large batch size allows the more effective use of GPUs. For the following performance numbers PP is prompt processing, and TG 128/1024 are the generation of 128/1024 tokens with an empty prompt:

| GPU      | Test    | Model    | t/s master | t/s PR | Speedup |
|----------|---------|----------|------------|--------|---------|
| RTX 3090 | PP      | 7b q4_0  |        452 |    990 |    2.19 |
| RTX 3090 | PP      | 13b q4_0 |        318 |    645 |    2.03 |
| RTX 3090 | PP      | 33b q4_0 |        155 |    292 |    1.88 |
| RTX 3090 | TG 128  | 7b q4_0  |      45.66 |  62.66 |    1.37 |
| RTX 3090 | TG 128  | 13b q4_0 |      29.94 |  38.21 |    1.28 |
| RTX 3090 | TG 128  | 33b q4_0 |      14.89 |  17.61 |    1.18 |
| RTX 3090 | TG 1024 | 7b q4_0  |      31.08 |  56.34 |    1.81 |
| RTX 3090 | TG 1024 | 13b q4_0 |      20.16 |  35.20 |    1.75 |
| RTX 3090 | TG 1024 | 33b q4_0 |      10.37 |  16.49 |    1.59 |

Note that I was only using a single thread for the PR since multiple threads have no benefit when all computations are on the GPU but they still add overhead.

I added CUDA kernels for scale, cpy, diag_mask_inf, and soft_max. I also added two special kernels for doing matrix vector multiplication with permuted or not contiguous inputs; they are used in conjunction with the KV cache.

Changes to `ggml.c`: I added a utility function `ggml_is_permuted`.

Things that are still to do:

* [x] Fix VRAM memory leaks.
* [x] Fix memory usage prints.
* [x] Check performance for lower-end GPUs and add a `--low-vram` option if necessary.
* [x] Check Windows performance and maybe disable features.
* [x] General code cleanup.