@KerfuffleV2 2023-06-12
> Check performance for lower-end GPUs and add a --low-vram option if necessary.

Unfortunately, I think that will be necessary. I have a 6GB 1060 - I can't even run `main` with 2048 context and `-ngl 0` on a 65B model. Using `-ngl 0 -c 1800` uses about 4.5GB VRAM even though the log output claims `total VRAM used: 512 MB`

Since I upgraded my CPU, offloading to the GPU isn't really worthwhile anymore except for prompt ingestion. It does make a big difference there so it would be unfortunate if just using cuBLAS prompt input was impossible for full context on larger models.