@JohannesGaessler 2023-06-12
@TheBloke I can reproduce the issue. It seems to have to do with -ngl 60; with those parameters some tensors are still on the CPU and the error happens when copying the data to the GPU. Setting -ngl 61 moves the non-repeating layers to the GPU as well and seems to work as intended.