@TheBloke 2023-06-12
Ahh OK, thanks. I did -ngl 60 because I thought that'd be the whole model.  Yes that fixes it, thank you.  Now tested fine up to 2000 tokens.

My first benchmark results are:
H100 + Intel(R) Xeon(R) Platinum 8480+:
* 7B q4_0 : main branch: 38.23 ms/token; PR: 33.49 ms/token (best was 31.38ms)
* 30B q4_0 : main branch: 100.33ms/token; PR: 85.61 ms/token (best was 79.28ms)