@JohannesGaessler 2023-06-12
Did you set `--threads 1`? You didn't do it in the command you posted previously and it should give a performance boost due to lower overhead.