@TheBloke 2023-06-12
Oh wow! My apologies, I missed that in your OP.

H100 + Intel(R) Xeon(R) Platinum 8480+.  500 tokens returned. 

* 7B q4_K_S :
  * main branch (`--threads 13`): 39.2ms/tok (25.51 tokens/s)
  * PR (`--threads 1`): 16.4 ms/token (60.97 tokens/s)
  * = 2.39x
* 30B q4_K_S:
  * main branch (`--threads 13`): 92.69 ms/token (10.79 tokens/s)
  * PR (`--threads 1`): 53.71 ms/token (18.62 tokens/s)
  * = 1.73x

This is really huge. This is the first time I've ever had GGML outperform GPTQ on any system.

The max performance I've ever had from GPTQ 7B is 98 tokens/s, but that requires both a fast GPU (4090) and a top single-core CPU (i9-13900K).   This system I'm on now has top GPU (H100), but middling single-core performing CPU, and it does around 45 tokens/s on 7B GPTQ 4bit.

I notice that llama.cpp is also pegged at 100% of one core, so I'm assuming a faster single-core CPU would likewise scale llama.cpp's figure.  So soon I'll test on that 4090 + i9-13900K system and see if llama.cpp can get close to that 100 t/s mark.

I just re-tested 30B GPTQ 4bit on this H100 system and got 20 t//s - still ahead of GGML, but only a tiny bit.

This is really game-changing. Well done!