@3dluvr 2023-06-19
That was a great tip @nauful - thank you !

Before:
```
llama_print_timings:        load time = 42949.69 ms
llama_print_timings:      sample time =   268.24 ms /   519 runs   (    0.52 ms per token)
llama_print_timings: prompt eval time =   324.68 ms /    17 tokens (   19.10 ms per token)
llama_print_timings:        eval time = 24496.90 ms /   518 runs   (   47.29 ms per token)
llama_print_timings:       total time = 25278.40 ms
```

After setting it to On:
```
llama_print_timings:        load time = 42962.13 ms
llama_print_timings:      sample time =   213.91 ms /   398 runs   (    0.54 ms per token)
llama_print_timings: prompt eval time =   462.09 ms /    17 tokens (   27.18 ms per token)
llama_print_timings:        eval time = 10362.89 ms /   397 runs   (   26.10 ms per token)
llama_print_timings:       total time = 11193.14 ms
```