@sirajperson 2023-06-23
This is why I've been fine tuning open_lama_3b for instruct:

```
./main -m ./models/alpaca_orca_open_llama_3b/ggml-model-f32.bin -ngl 10000 --no-mmap -p " \
Below is an instruction that describes a task. Write a response that appropriately completes the request
### Instruction: write a fibb function in python
### Response:

"

main: build = 725 (7487137)
main: seed  = 1687505508
ggml_init_cublas: found 1 CUDA devices:
  Device 0: NVIDIA H100 PCIe
llama.cpp: loading model from ./models/alpaca_orca_open_llama_3b/ggml-model-f32.bin
llama_model_load_internal: format     = ggjt v1 (pre #1405)
llama_model_load_internal: n_vocab    = 32000
llama_model_load_internal: n_ctx      = 512
llama_model_load_internal: n_embd     = 3200
llama_model_load_internal: n_mult     = 240
llama_model_load_internal: n_head     = 32
llama_model_load_internal: n_layer    = 26
llama_model_load_internal: n_rot      = 100
llama_model_load_internal: ftype      = 0 (all F32)
llama_model_load_internal: n_ff       = 8640
llama_model_load_internal: n_parts    = 1
llama_model_load_internal: model size = 3B
llama_model_load_internal: ggml ctx size = 13071.02 MB
llama_model_load_internal: using CUDA for GPU acceleration
llama_model_load_internal: mem required  = 1414.68 MB (+  682.00 MB per state)
llama_model_load_internal: allocating batch_size x 1 MB = 512 MB VRAM for the scratch buffer
llama_model_load_internal: offloading 26 repeating layers to GPU
llama_model_load_internal: offloading non-repeating layers to GPU
llama_model_load_internal: offloading v cache to GPU
llama_model_load_internal: offloading k cache to GPU
llama_model_load_internal: offloaded 29/29 layers to GPU
llama_model_load_internal: total VRAM used: 13875 MB
.................................................................................................
llama_init_from_file: kv self size  =  162.50 MB

system_info: n_threads = 13 / 26 | AVX = 1 | AVX2 = 1 | AVX512 = 1 | AVX512_VBMI = 1 | AVX512_VNNI = 1 | FMA = 1 | NEON = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 1 | VSX = 0 | 
sampling: repeat_last_n = 64, repeat_penalty = 1.100000, presence_penalty = 0.000000, frequency_penalty = 0.000000, top_k = 40, tfs_z = 1.000000, top_p = 0.950000, typical_p = 1.000000, temp = 0.800000, mirostat = 0, mirostat_lr = 0.100000, mirostat_ent = 5.000000
generate: n_ctx = 512, n_batch = 512, n_predict = -1, n_keep = 0


  Below is an instruction that describes a task. Write a response that appropriately completes the request
### Instruction: write a fibb function in python
### Response:

Here's an example of a fibonacci function in Python:

'''python
def fibonacci(n):
    return list(map(lub, range(2, n+1)))
'''

This function takes an integer `n` as input and returns a list of the first `n` numbers of the Fibonacci sequence:

[0, 1, 1, 2, 3, 5, 8, 13, 21, 34]

You can call this function with any number as input, and it will return the first `n` numbers of the Fibonacci sequence. [end of text]

llama_print_timings:        load time = 10925.31 ms
llama_print_timings:      sample time =    65.53 ms /   158 runs   (    0.41 ms per token,  2411.11 tokens per second)
llama_print_timings: prompt eval time =    55.92 ms /    38 tokens (    1.47 ms per token,   679.49 tokens per second)
llama_print_timings:        eval time =  2307.79 ms /   157 runs   (   14.70 ms per token,    68.03 tokens per second)
```
Look at those timings :-D